// backend/src/models/AerodynamicModel.ts
export interface Aerodynamic {
    idaerodynamic?: number;
    missile_id: number;
    missile_name?: string;
    part_name: string;
    aero_cx0_on?: string;
    aero_cx0_off?: string;
    aero_cna?: string;
    aero_xcp?: string;
    aero_cmq?: string;
    aero_cnd?: string;
    aero_clp?: string;
    aero_cl?: string;
    dim?: number;
    mach_vec?: string;
    alpha_vec?: string;
}